+++ packages/assert/src/assert.ts
@@ -6,9 +6,9 @@
  */
 
 import * as axe from 'axe-core';
-import { A11yConfig, extended } from '@sa11y/preset-rules';
+import { A11yConfig, recommended } from '@sa11y/preset-rules';
 import { A11yError } from '@sa11y/format';
-import { axeRuntimeExceptionMsgPrefix } from '@sa11y/common';
+import { getViolations } from '@sa11y/common';
 
 /**
  * Type def for context that can be checked for accessibility.
@@ -19,19 +19,15 @@ export type a11yCheckableContext = Document | HTMLElement;
 /**
  * Checks DOM for accessibility issues and throws an error if violations are found.
  * @param context - DOM or HTMLElement to be tested for accessibility
- * @param rules - A11yConfig preset rule to use, defaults to extended
+ * @param rules - A11yConfig preset rule to use, defaults to recommended
  * @throws error - with the accessibility issues found, does not return any value
  * */
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export async function assertAccessible(context: a11yCheckableContext = document, rules: A11yConfig = extended) {
-    let violations;
-    try {
+export async function assertAccessible(context: a11yCheckableContext = document, rules: A11yConfig = recommended) {
+    const violations = await getViolations(async () => {
         const results = await axe.run(context as axe.ElementContext, rules as axe.RunOptions);
-        violations = results.violations;
-    } catch (e) {
-        throw new Error(`${axeRuntimeExceptionMsgPrefix} ${e}`);
-    }
-    if (violations.length > 0) {
-        throw new A11yError(violations);
-    }
+        return results.violations;
+    });
+
+    A11yError.checkAndThrow(violations);
 }