+++ packages/format/src/format.ts
@@ -70,16 +70,18 @@ export function sortViolations(violations: Result[]): void {
  *  Custom error object to represent a11y violations
  */
 export class A11yError extends Error {
+    static readonly errMsgHeader = 'accessibility issues found';
+
     constructor(readonly violations: Result[]) {
-        super(`${violations.length} accessibility issues found`);
+        super(`${violations.length} ${A11yError.errMsgHeader}`);
         this.name = A11yError.name;
+        this.message = `${violations.length} ${A11yError.errMsgHeader}\n ${this.format()}`;
     }
 
-    get message(): string {
-        // TODO (debug): Why is this not used (in code cov) even when A11yError.message is called ?
-        //  Looks like the super().message is invoked
-        /* istanbul ignore next */
-        return this.format();
+    static checkAndThrow(violations: Result[]): void {
+        if (violations.length > 0) {
+            throw new A11yError(violations);
+        }
     }
 
     get length(): number {
@@ -92,7 +94,7 @@ export class A11yError extends Error {
      */
     format(options: Options = {}): string {
         options = { ...DefaultOptions, ...options };
-        if (options.formatter) {
+        if (options.formatter !== undefined) {
             return options.formatter(this.violations);
         }