+++ packages/test-utils/src/test-data.ts
@@ -5,33 +5,17 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export const domWithA11yIssuesBodyID = 'dom-with-issues';
-// DOM with no a11y issues
-export const domWithA11yIssues = `<html>
-                            <body>
-                             <div id="${domWithA11yIssuesBodyID}">
-                                <a href="#"></a>
-                             </div>
-                            </body>
-                           </html>`;
+import fs from 'fs';
+import path from 'path';
 
+// DOM with a11y issues
+export const domWithA11yIssuesBodyID = 'dom-with-issues';
+const fileWithA11yIssues = path.resolve(__dirname, '../__data__/a11yIssues.html');
+export const htmlFileWithA11yIssues = 'file:///' + fileWithA11yIssues;
+export const domWithA11yIssues = fs.readFileSync(fileWithA11yIssues).toString();
+
+// DOM containing no a11y issues
 export const shadowDomID = 'upside-down';
-// DOM containing a11y issues
-// From https://github.com/dequelabs/axe-selenium-java/blob/develop/src/test/resources/test-app.js-->
-export const domWithNoA11yIssues = `<!doctype html>
-                            <html lang="en">
-                            <head>
-                                <title>Test Page</title>
-                            </head>
-                            <body>
-                            <div role="main" id="host">
-                                <h1>This is a test</h1>
-                                <p>This is a test page with no violations</p>
-                            </div>
-                            <div role="contentinfo" id=${shadowDomID}></div> <!-- cSpell:disable-line -->
-                                <script>
-                                    const shadow = document.getElementById("upside-down").attachShadow({mode: "open"});
-                                    shadow.innerHTML = '<h2 id="shadow">SHADOW DOM</h2><ul><li>Shadow Item 1</li></ul>'
-                                </script>
-                            </body>
-                            </html>`;
+const fileWithNoA11yIssues = path.resolve(__dirname, '../__data__/noA11yIssues.html');
+export const htmlFileWithNoA11yIssues = 'file:///' + fileWithNoA11yIssues;
+export const domWithNoA11yIssues = fs.readFileSync(fileWithNoA11yIssues).toString();